PyTableExtraction/requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pandas>=2.0.0
matplotlib>=3.7.0
torch>=2.0.0
torchvision>=0.15.0
easyocr>=1.7.0
pytesseract>=0.3.10
scikit-learn>=1.3.0 